requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
packaging==21.3
pdf2image==1.16.0
Pillow==9.0.1
pyparsing==3.0.8
pytesseract==0.3.9
requests==2.27.1
urllib3==1.26.8
